[#46] Use Alpine instead of Debian in testing
Use Alpine Python images in docker-robottests instead of Debian. Debian
is overkill as basically only SSH server and client is needed along with
Python.

Bash is installed as the default shell of the RemoteRunner is bash.
Moreover, RemoteScript tests uses SFTP in file transfers.  RemoteScript
is tested along with RemoteRunner to verify the feature parity.